Extract shared mission phase reference helpers
