+++ apps/desktop/src/main/services/orchestrator/coordinatorAgent.ts
@@ -72,6 +72,7 @@ import {
   normalizeCoordinatorToolName,
   updateDelegationContract,
 } from "./delegationContracts";
+import { findPhaseByRef, resolveMustPrecedePredecessors } from "../missions/phaseEngine";
 
 // ---------------------------------------------------------------------------
 // Types
@@ -2584,22 +2585,6 @@ export class CoordinatorAgent {
       .some((step) => !TERMINAL_STEP_STATUSES.has(step.status));
   }
 
-  private phaseMatchesRef(phase: PhaseCard, ref: string): boolean {
-    const normalizedRef = ref.trim().toLowerCase();
-    if (!normalizedRef) return false;
-    return phase.phaseKey.trim().toLowerCase() === normalizedRef || phase.name.trim().toLowerCase() === normalizedRef;
-  }
-
-  private findPhaseByRef(phases: PhaseCard[], ref: string): PhaseCard | null {
-    return phases.find((phase) => this.phaseMatchesRef(phase, ref)) ?? null;
-  }
-
-  private resolveMustPrecedePredecessors(phases: PhaseCard[], targetPhase: PhaseCard): PhaseCard[] {
-    return phases.filter((phase) =>
-      (phase.orderingConstraints.mustPrecede ?? []).some((successorRef) => this.phaseMatchesRef(targetPhase, successorRef))
-    );
-  }
-
   private resolvePhaseDependencyStepKeys(
     graph: ReturnType<CoordinatorAgentDeps["orchestratorService"]["getRunGraph"]>,
     phase: PhaseCard,
@@ -2612,12 +2597,12 @@ export class CoordinatorAgent {
     const predecessorPhases =
       explicitMustFollow.length > 0
         ? explicitMustFollow
-            .map((key) => this.findPhaseByRef(sorted, key))
+            .map((key) => findPhaseByRef(sorted, key))
             .filter((candidate): candidate is PhaseCard => Boolean(candidate))
         : sorted.slice(0, Math.max(0, targetIndex)).filter((candidate) =>
             candidate.orderingConstraints.mustBeFirst || candidate.validationGate.required === true
           );
-    for (const predecessor of this.resolveMustPrecedePredecessors(sorted, phase)) {
+    for (const predecessor of resolveMustPrecedePredecessors(sorted, phase)) {
       if (!predecessorPhases.includes(predecessor)) predecessorPhases.push(predecessor);
     }
     const dependencyKeys: string[] = [];
@@ -2653,10 +2638,10 @@ export class CoordinatorAgent {
       if (phase.orderingConstraints.mustBeLast && this.phaseHasOpenExecutionStepInGraph(graph, earlier)) return false;
     }
     for (const predecessorKey of explicitMustFollow) {
-      const predecessor = this.findPhaseByRef(sorted, predecessorKey);
+      const predecessor = findPhaseByRef(sorted, predecessorKey);
       if (predecessor && !this.phaseHasValidatedSuccess(graph, predecessor)) return false;
     }
-    for (const predecessor of this.resolveMustPrecedePredecessors(sorted, phase)) {
+    for (const predecessor of resolveMustPrecedePredecessors(sorted, phase)) {
       if (!this.phaseHasValidatedSuccess(graph, predecessor)) return false;
     }
     return true;
